requirements/base.txt
argon2-cffi
whitenoise
psycopg2-binary
pillow

# Django
# ------------------------------------------------------------------------------
django
django-environ
django-jazzmin
# Django REST Framework
djangorestframework
django-cors-headers
djangorestframework-simplejwt
djoser
django-filter
drf-spectacular
